+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.7 (2016-06-30)
+Features:
+  - (For Debugging Only) Long-click to force testing and labeling all elements.
+
+Bugfixes:
+  - Fix the performance issue of `addSpaces()`.
+
 ## 1.0.5 (2016-06-29)
 Bugfixes:
   - Hidden elements can be deteced now, even if they are hidden inside an inline element.